+++ devicemodel/Makefile
@@ -106,7 +106,7 @@ SRCS += hw/platform/ioc.c
 SRCS += hw/platform/ioc_cbc.c
 SRCS += hw/platform/pty_vuart.c
 SRCS += hw/platform/acpi/acpi.c
-SRCS += hw/platform/acpi/rtct.c
+SRCS += hw/platform/vssram/vssram.c
 SRCS += hw/platform/acpi/acpi_pm.c
 SRCS += hw/platform/rpmb/rpmb_sim.c
 SRCS += hw/platform/rpmb/rpmb_backend.c
